+++ ReleaseNotes.md
@@ -4,7 +4,7 @@
 
 - Added InputRotatesSprite and VelocityRotatesSprite to Flip Behaviour on platformer controller. Rotating instead of scaling changes the actual objects direction, so it can affect other movement systems, for example.
 - Movement components can now have conditions built in which take care of activating/deactivating that movement
-
+- Added Dash action, which enable short bursts of movement, using variables or values
 
 ## V1.13.0